requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/altgraph-0.17.2-py2.py3-none-any.whl
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
distro==1.8.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/future-0.18.2-py3-none-any.whl
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
openai==1.3.5
pdf2image==1.16.3
Pillow==10.1.0
prompt-toolkit==3.0.41
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
redis==5.0.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot11/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-141/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
zipp==3.17.0
